zsh: Remove custom key bindings C-f and C-r
C-r is already bound to history-incremental-search-backward and C-f is
redundant with C-s.

Also, make C-q and C-s work in zsh. Those keys are traditionally used
for software control flow purposes, and it is not something I have ever
used (on purpose anyway). Having C-s (and C-q) available in zsh is far
more useful.